Moved two function belonging to dkgRetryLoop to dkg_loop.go
These two functions were mistakenly omitted in the refactoring performed
earlier. Moving them now!
